+++ tests/local/test-map.R
@@ -28,29 +28,3 @@ for (index in seq_len(3L)) {
   )
 }
 controller$terminate()
-
-# In case the above gets stuck again, work on the reprex below.
-# nolint start
-#
-# library(mirai)
-# library(nanonext)
-# library(purrr)
-# daemons(n = 4, url = "ws://127.0.0.1:5700")
-# n_tasks <- 20000
-# tasks <- replicate(n_tasks, mirai(TRUE))
-# cv1 <- nextget("cv")
-# cv2 <- cv()
-# cv1 %~>% cv2
-# walk(rownames(status()$daemons), launch_local)
-# while(cv_value(cv1) < n_tasks) {
-#   until(cv2, msec = 500)
-# }
-# all(!map_lgl(tasks, unresolved))
-# tasks <- replicate(n_tasks, mirai(TRUE))
-# while(cv_value(cv1) < 2 * n_tasks) {
-#   until(cv2, msec = 500)
-# }
-# all(!map_lgl(tasks, unresolved))
-# daemons(n = 0)
-#
-# nolint end